+++ react/typings/global.d.ts
@@ -47,7 +47,7 @@ interface OrderItem {
   skuName: string
   name: string
   productId: string
-  price: number
+  sellingPrice: number
   listPrice: number
   bundleItems: Bundle[]
   isGift: boolean